Add offer_categorisation_model
